Filter Refacturing
